update `EnvUtils.getJavaVersion` to use JVM system property
