+++ maestro-cli/src/main/java/maestro/cli/update/Updates.kt
@@ -17,21 +17,6 @@ import kotlin.io.path.readText
 import kotlin.io.path.writeText
 
 object Updates {
-    val JAVA_VERSION: Int
-        get() {
-            // Adapted from https://stackoverflow.com/a/2591122/7009800
-            val version = System.getProperty("java.version")
-            return if (version.startsWith("1.")) {
-                version.substring(2, 3).toInt()
-            } else {
-                val dot = version.indexOf(".")
-                if (dot != -1) version.substring(0, dot).toInt() else 0
-            }
-        }
-
-    val XCODE: String
-        get() = EnvUtils.getXcodeVersion()
-
     val OS_NAME: String = System.getProperty("os.name")
     val OS_ARCH: String = System.getProperty("os.arch")
     val DEVICE_UUID: String
